Moved crds.tables module to crds.io.tables and adjusted module imports and unit tests accordingly.
