+++ crds/bestrefs/table_effects.py
@@ -15,7 +15,7 @@ from __future__ import absolute_import
 
 from crds.core import rmap, log
 from crds.core.python23 import *
-from crds import tables
+from crds.io import tables
 from crds.client import api
 
 def is_reprocessing_required(dataset,  dataset_parameters, old_context, new_context, update):
