+++ crds/certify/validators.py
@@ -22,7 +22,7 @@ from crds.core import log, config, utils, timestamp, selectors
 from crds.core.exceptions import MissingKeywordError, IllegalKeywordError
 from crds.core.exceptions import TpnDefinitionError, RequiredConditionError
 from crds.core.exceptions import BadKernelSumError
-from crds import tables
+from crds.io import tables
 from crds import data_file
 
 from . import reftypes
